requirements.txt
absl-py==1.0.0
attrs==21.4.0
cachetools==5.2.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
google-api-core==2.8.2
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
importlab==0.7
importlib-resources==5.4.0
jsonschema==4.4.0
libcst==0.4.1
mailjet-rest==1.3.4
mypy-extensions==0.4.3
networkx==2.6.3
ninja==1.10.2.3
notifiers==1.3.0
oauthlib==3.2.0
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyjokes==0.6.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.21.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tabulate==0.8.9
toml==0.10.2
typing-inspect==0.7.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
yattag==1.14.0
zipp==3.7.0
